+++ pyscf/dh/test/test_dfdh_dipole.py
@@ -26,20 +26,20 @@ class TestDFDipole:
         mol = _mol()
         dip_nuc = np.einsum("At, A-> t", mol.atom_coords(), mol.atom_charges())
         nde = NumericDiff(DipoleDerivGenerator(_mol_to_eng(mol, "B2PLYP"))).derivative + dip_nuc
-        de = DFDH(mol, "B2PLYP").run().dipole()
+        de = DFDH(mol, "B2PLYP").run().polar_method().dipole()
         assert np.allclose(nde, de, atol=1e-5, rtol=1e-4)
 
     def test_mp2(self):
         REF = np.array([0.513968430874, 0.494934113883, 0.469052040835])
-        de = DFDH(_mol(), "MP2").run().dipole()
+        de = DFDH(_mol(), "MP2").run().polar_method().dipole()
         assert np.allclose(REF, de, atol=1e-5, rtol=1e-4)
 
     def test_xyg3(self):
         REF = np.array([0.516715680877, 0.497164080305, 0.467699209941])
-        de = DFDH(_mol(), "XYG3").run().dipole()
+        de = DFDH(_mol(), "XYG3").run().polar_method().dipole()
         assert np.allclose(REF, de, atol=1e-5, rtol=1e-4)
 
     def test_xygj_os(self):
         REF = np.array([0.519030495499, 0.498740432488, 0.468021122351])
-        de = DFDH(_mol(), "XYGJ-OS").run().dipole()
+        de = DFDH(_mol(), "XYGJ-OS").run().polar_method().dipole()
         assert np.allclose(REF, de, atol=1e-5, rtol=1e-4)
